feat: mark progress in tx_seen_in_blocks in IFEs to prevent re-searching
feat: improve the previous commit - search for each ife in block once

if search is undertaken, it means that the input was spent. If it was spent, but the tx was not found it means it was double-spent wrt the IFE tx

Hence the IFE tx will never get to the valid blocks. No need to look again ever